% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_between_curves.R
\name{convert_between_curves}
\alias{convert_between_curves}
\title{Convert between curves.}
\usage{
convert_between_curves(mfi, params_new, params_ref_new, params_ref_old)
}
\arguments{
\item{mfi}{Known mfi of samples to be converted}

\item{params_new}{Known parameters for five parameter logistic fit.}

\item{params_ref_new}{Known parameters for five parameter logistic fit on reference plate}

\item{params_ref_old}{Known parameters for five parameter logistic fit of the old beads on the sample plate.}
}
\value{
Returns the predicted dilution in comparison to the reference plate
}
\description{
Convert mfi on new plate to the dilution on the reference plate.
}
