% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_mfi_to_dilution.R
\name{convert_mfi_to_dilution}
\alias{convert_mfi_to_dilution}
\title{Convert mfi to dilution using known standard curve fit.}
\usage{
convert_mfi_to_dilution(mfi, params)
}
\arguments{
\item{mfi}{Known mfi of samples}

\item{params}{Known parameters for five parameter logistic fit.}
}
\value{
Returns the dilution of each sample in mfi.
}
\description{
Convert mfi to dilution using known standard curve fit.
}
